+++ NEWS.md
@@ -1,3 +1,5 @@
+# modeldata (development version)
+
 # modeldata 1.0.1
 
 * Small update to fix HTML  for CRAN.